+++ CHANGELOG.md
@@ -14,6 +14,7 @@ you need to hover again to see it.
 
 - Changed internal environment to be more decoupled 
 - Fixed code highlight for methods
+- Fixed bad package when package does not exist
 
 ## 0.1.1 - 230115